requirements.txt
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mmcv-full==1.3.0
mmdet==2.11.0
mmpycocotools==12.0.3
numpy==1.21.5
onnx==1.11.0
opencv-contrib-python-headless
opencv-python
opencv-python-headless
Pillow==9.1.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.5.4
torchcontrib==0.0.2
tqdm==4.64.0